refactor(web): split useKimiWebClient into workspace/sideChat/modelProvider modules
